upd: Coloca o cr_executor_correctio dentro do cr_executor
